requirements.txt
Django==5.2.12
djangorestframework>=3.14.0,<3.15
djangorestframework-simplejwt>=5.3.1,<5.4
django-cors-headers>=4.3.1,<5.0

# LangChain & AI (CPU-optimized)
langchain>=0.1.0,<0.2
langchain-community>=0.0.10,<0.1
langchain-huggingface>=0.0.1,<0.1

# Use smaller, faster embedding model for CPU
sentence-transformers>=2.6.0,<3.0

# Vector Database (CPU)
faiss-cpu>=1.7.4,<2.0

# File Handling
pypdf>=4.0.0,<5.0

# Async Tasks
celery>=5.3.4,<5.4
redis>=5.0.1,<6.0

# Utilities
python-dotenv>=1.0.0,<2.0
waitress>=3.0.0,<4.0
whitenoise>=6.6.0,<7.0

# [CPU-Only PyTorch] - Install separately in Dockerfile
# torch>=2.0.0 will be installed with --extra-index-url for CPU